Refactor variable types and fix bug in ReadRulerList
Changed variable types in VolumeRenderer.cpp and tif_reader.cpp to size_t and uint64_t respectively for better handling of large values. Removed unused variable szval and corrected the variable used in SetWorkTime call in Project.cpp to fix a bug.

+++ FluoRender/FluoRender/Flrender/VolumeRenderer.cpp
@@ -1728,8 +1728,8 @@ namespace flvr
 		if (c<0 || c>=TEXTURE_MAX_COMPONENTS)
 			return;
 
-		int i;
-		unsigned int num = static_cast<unsigned int>(bricks->size());
+		size_t i;
+		size_t num = bricks->size();
 		for (i=((order==2)?(num-1):0);
 			(order==2)?(i>=0):(i<num);
 			i+=((order==2)?-1:1))
